fix: revert to minimum when amount over max
Solved a bug where if a user typed in the text field of the spinner an amount greater than what is allowed, they could order but not at the maximum they set but another value instead. This new addition warns them that they breached the max and reverts the value back to the minimum.